MAT-8929: Preventing API updates when a lock exists on a Measure by a different user
