+++ src/main/java/cms/gov/madie/measure/services/MeasureService.java
@@ -11,9 +11,9 @@ import gov.cms.madie.models.common.*;
 import gov.cms.madie.models.dto.LibraryUsage;
 import gov.cms.madie.models.measure.*;
 import jakarta.annotation.Nullable;
-import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -27,8 +27,7 @@ import java.util.stream.Collectors;
 
 @Slf4j
 @Service
-@AllArgsConstructor
-public class MeasureService {
+public class MeasureService extends BaseMeasureService {
   private final MeasureRepository measureRepository;
   private final MeasureSetRepository measureSetRepository;
   private final TestCasePatchRepository testCasePatchRepository;
@@ -37,10 +36,38 @@ public class MeasureService {
   private final ActionLogService actionLogService;
   private final MeasureSetService measureSetService;
   private final CqlTemplateConfigService cqlTemplateConfigService;
-
   private final TerminologyValidationService terminologyValidationService;
   private final AppConfigService appConfigService;
   private final MeasureLockService measureLockService;
+
+  @Autowired
+  public MeasureService(
+      MeasureRepository measureRepository,
+      MeasureSetRepository measureSetRepository,
+      TestCasePatchRepository testCasePatchRepository,
+      ElmTranslatorClient elmTranslatorClient,
+      MeasureUtil measureUtil,
+      ActionLogService actionLogService,
+      MeasureSetService measureSetService,
+      CqlTemplateConfigService cqlTemplateConfigService,
+      TerminologyValidationService terminologyValidationService,
+      AppConfigService appConfigService,
+      MeasureLockService measureLockService) {
+    // Pass parent dependencies to BaseMeasureService constructor
+    super(measureRepository, measureSetService, appConfigService, measureLockService);
+    // Assign child-specific fields
+    this.measureRepository = measureRepository;
+    this.measureSetRepository = measureSetRepository;
+    this.testCasePatchRepository = testCasePatchRepository;
+    this.elmTranslatorClient = elmTranslatorClient;
+    this.measureUtil = measureUtil;
+    this.actionLogService = actionLogService;
+    this.measureSetService = measureSetService;
+    this.cqlTemplateConfigService = cqlTemplateConfigService;
+    this.terminologyValidationService = terminologyValidationService;
+    this.appConfigService = appConfigService;
+    this.measureLockService = measureLockService;
+  }
 
   public void verifyAuthorizationByMeasureSetId(
       String username, String measureSetId, boolean ownerOnly) {
@@ -64,59 +91,6 @@ public class MeasureService {
    */
   public void verifyAuthorization(String username, Measure measure) {
     verifyAuthorization(username, measure, List.of(RoleEnum.SHARED_WITH));
-  }
-
-  /**
-   * Verifies the specified user has privileges on the given measure based on measure owner and the
-   * passed roles. Providing null or empty roles will perform an authorization check for owner only.
-   *
-   * @param username
-   * @param measure
-   * @param roles
-   */
-  public void verifyAuthorization(String username, Measure measure, List<RoleEnum> roles) {
-    MeasureSet measureSet =
-        measure.getMeasureSet() == null
-            ? measureSetService.findByMeasureSetId(measure.getMeasureSetId())
-            : measure.getMeasureSet();
-    if (measureSet == null) {
-      log.error(
-          "User [{}] called verifyAuthorization but failed because no measure set exists for "
-              + "measure with measure ID [{}]",
-          username,
-          measure.getId());
-      throw new InvalidMeasureStateException(
-          "No measure set exists for measure with ID " + measure.getId());
-    }
-    MeasureServiceUtil.verifyMeasureSetAuthorization(
-        username, "Measure", measure.getId(), roles, measureSet);
-  }
-
-  // TODO: start replacing usage of measureRepository.findById with this method
-  public Measure findMeasureById(final String id) {
-    return measureRepository
-        .findById(id)
-        .map(
-            m -> {
-              Measure.MeasureBuilder builder =
-                  m.toBuilder()
-                      .measureSet(measureSetService.findByMeasureSetId(m.getMeasureSetId()));
-
-              // Map measure lock if locking feature is enabled
-              if (appConfigService.isFlagEnabled(MadieFeatureFlag.LOCKING)) {
-                MeasureLock lock = measureLockService.findByMeasureId(m.getId());
-                if (lock != null) {
-                  builder.measureLock(
-                      gov.cms.madie.models.measure.MeasureLock.builder()
-                          .id(lock.getId())
-                          .lockedBy(lock.getLockedBy())
-                          .build());
-                }
-              }
-
-              return builder.build();
-            })
-        .orElse(null);
   }
 
   public Measure findActiveMeasureById(@Nullable String measureId) {
@@ -299,9 +273,7 @@ public class MeasureService {
       try {
         outputMeasure =
             measureUtil.validateAllMeasureDependencies(updateElm(updatingMeasure, accessToken));
-
-        // remove this condition when we validate for terminology service errors in
-        // backend
+        // remove this condition when we validate for terminology service errors in backend
         if (!outputMeasure.isCqlErrors()) {
           outputMeasure.setCqlErrors(updatingMeasure.isCqlErrors());
         }
@@ -574,68 +546,11 @@ public class MeasureService {
     return measureIdToAclSpecification;
   }
 
-  /**
-   * Verifies that the user is authorized to perform share/unshare operations on the given measures.
-   * - Restrict sharing to owners only (ownerOnly = true). - Allow unsharing by owners or users who
-   * the measure is already shared with (ownerOnly = false).
-   */
-  private void verifyShareAuthorization(
-      Map<String, List<String>> measureUserIdMap, String username, boolean ownerOnly) {
-    log.info(
-        "User [{}] has called verifyShareAuthorization to determine whether operation with [{}]"
-            + " is allowed to be performed",
-        username,
-        measureUserIdMap);
-
-    measureUserIdMap
-        .keySet()
-        .forEach(
-            measureId -> {
-              Measure measure = findMeasureById(measureId);
-
-              if (measure == null) {
-                log.error(
-                    "User [{}] called verifyShareAuthorization with measureUserIdMap [{}] but "
-                        + "failed because the measure with measure ID [{}] does not exist.",
-                    username,
-                    measureUserIdMap,
-                    measureId);
-                throw new ResourceNotFoundException("Measure does not exist: " + measureId);
-              }
-              verifyAuthorization(
-                  username, measure, ownerOnly ? List.of() : List.of(RoleEnum.SHARED_WITH));
-            });
-
-    log.info(
-        "User [{}] successfully called verifyShareAuthorization and determined that operation "
-            + "with [{}] is allowed to be performed",
-        username,
-        measureUserIdMap);
-  }
-
-  private AclOperation buildShareAclOperation(List<String> userIds) {
-    return AclOperation.builder()
-        .acls(buildShareAclSpecifications(userIds))
-        .action(AclOperation.AclAction.GRANT)
-        .build();
-  }
-
   private AclOperation buildUnshareAclOperation(List<String> userIds) {
     return AclOperation.builder()
         .acls(buildShareAclSpecifications(userIds))
         .action(AclOperation.AclAction.REVOKE)
         .build();
-  }
-
-  private List<AclSpecification> buildShareAclSpecifications(List<String> userIds) {
-    return userIds.stream()
-        .map(
-            userId ->
-                AclSpecification.builder()
-                    .userId(userId)
-                    .roles(Set.of(RoleEnum.SHARED_WITH))
-                    .build())
-        .toList();
   }
 
   public void changeOwnership(
